Use nested sequences for discriminabors
With the changes in Daffodil for DAFFODIL-1971 we need to use nested
sequences to maintain the same discriminator as previously. In this
case, the discriminators were simply using fn:true(), which prior to
DAFFODIL-1971 would only be executed if the rest of the body
successfully parsed. After DAFFODIL-1971 all discriminator expressions
are attempted regardless if the rest of the sequence body parses.